# api/openapi.yaml
openapi: 3.1.0
info:
  title: PeerGit Platform API
  version: 0.1.0
  description: Health endpoints, identity and campus administration, media, and Phase 3 project/team/recruitment workflows.
servers:
  - url: https://peergit.localhost
paths:
...
  /api/v1/auth/github:
    get:
      operationId: startGitHubSignIn
      summary: Start a state- and S256-PKCE-protected GitHub App user authorization flow
      parameters:
        - in: query
          name: invite
          schema: { type: string, maxLength: 128 }
          description: Optional single-use invitation token.
      responses:
        '302': { description: Redirect to the identity provider }
        '429': { description: Sign-in rate limit reached }
        '503': { description: GitHub sign-in is not configured or provider is unavailable }
  /api/v1/auth/github/callback:
    get:
      operationId: finishGitHubSignIn
      summary: Resolve the immutable GitHub user ID and verified primary email and establish a limited opaque session
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: state
          required: true
          schema: { type: string }
      responses:
        '303': { description: Sign-in completed; session cookie set }
        '401': { description: Invalid, expired, replayed, or unverified sign-in }
        '403': { description: GitHub account lacks an authenticated verified primary email }
...
  /api/v1/me:
    get:
      operationId: currentProfile
      summary: Read the current account profile, skills, roles, and consent state
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Profile data }
        '401': { description: Authentication required }
  /api/v1/me/profile:
    patch:
      operationId: updateProfile
      summary: Update profile fields and self-reported skills
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdate' }
      responses:
        '200': { description: Profile updated }
        '400': { description: Invalid profile fields }
        '403': { description: CSRF or media-ownership check failed }
  /api/v1/me/consent:
    put:
      operationId: updateConsent
      summary: Grant or revoke versioned purpose-specific consent
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsentUpdate' }
      responses:
        '200': { description: Consent updated }
        '400': { description: Invalid or unsupported policy version }
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current server-side session
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: Session revoked and cookie cleared }
  /api/v1/auth/mfa/enroll:
    post:
      operationId: enrollMfa
      summary: Begin or replace the current user's TOTP setup
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200': { description: One-time TOTP setup secret }
        '403': { description: Existing MFA must be freshly verified before replacement }
  /api/v1/auth/mfa/confirm:
    post:
      operationId: confirmMfa
      summary: Confirm TOTP setup and issue one-time recovery codes
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MfaCode' }
      responses:
        '200': { description: MFA enabled; recovery codes shown once }
        '401': { description: Code invalid }
  /api/v1/auth/mfa/verify:
    post:
      operationId: verifyMfa
      summary: Elevate the current session for ten minutes using TOTP or a recovery code
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MfaCode' }
      responses:
        '200': { description: MFA verified }
        '401': { description: Code invalid }
...
components:
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-peergit_session
  schemas:
...
    ProfileUpdate:
      type: object
      required: [display_name]
      properties:
        display_name: { type: string, minLength: 1, maxLength: 120 }
        headline: { type: string, maxLength: 160 }
        bio: { type: string, maxLength: 2000 }
        avatar_media_id: { type: string, format: uuid }
        skills: { type: array, maxItems: 30, items: { type: string, minLength: 2, maxLength: 60 } }
    ConsentUpdate:
      type: object
      required: [purpose, policy_version, granted]
      properties:
        purpose: { type: string, enum: [terms, privacy, profile_discovery, portfolio, recruiter_discovery, analytics] }
        policy_version: { type: string, maxLength: 40 }
        granted: { type: boolean }
    MfaCode:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 6, maxLength: 32 }
...